Rely on MongoDB's ObjectId for ordering of events
Having a monotonically increasing ID is not as easy as in a relational
database as MongoDB doesn't have such IDs.

Order guarantees are not as good for ObjectIds, as for real
monotonically increasing IDs but it might be sufficient for this use
case as we don't expect more than one operation per second per aggregate.

More information about IDs in MongoDB can be found here:
https://www.mongodb.com/blog/post/generating-globally-unique-identifiers-for-use-with-mongodb

More insights about ordering guarantees can be found here:
https://www.mongodb.com/community/forums/t/objectid-guarantees-absolute-insertion-order-if-just-one-replica-set-and-no-sharding-throughout-lifetime/16902